+++ examples/twilio_stream_example.py
@@ -1,14 +1,82 @@
 import argparse
+import asyncio
 import base64
 import json
 import logging
+from typing import AsyncGenerator
 
 import aiohttp.web
+import g711
+import numpy
 
-from fixie_sdk.voice import audio_phone
+from fixie_sdk.voice import audio_base
 from fixie_sdk.voice import types
 from fixie_sdk.voice.session import VoiceSession
 from fixie_sdk.voice.session import VoiceSessionParams
+
+
+class PhoneAudioSink(audio_base.AudioSink):
+    """AudioSink that plays to the phone stream."""
+
+    def __init__(self, ws: aiohttp.web.WebSocketResponse) -> None:
+        super().__init__()
+        self._ws = ws
+        self._stream_sid = ""
+        self._sequence_no = 0
+
+    async def start(self, sample_rate: int = 8000, num_channels: int = 1):
+        pass
+
+    @property
+    def stream_sid(self):
+        return self._stream_sid
+
+    @stream_sid.setter
+    def stream_sid(self, value):
+        self._stream_sid = value
+
+    async def write(self, chunk: bytes) -> None:
+        audio_data = numpy.frombuffer(chunk, dtype=numpy.int16).astype(numpy.float32)
+        ulaw = g711.encode_ulaw(audio_data)
+        pay_load = base64.b64encode(ulaw)
+        # Send media message
+        media_data = {
+            "event": "media",
+            "streamSid": self._stream_sid,
+            "media": {"payload": pay_load},
+        }
+        media = json.dumps(media_data)
+        await self._ws.send_str(media)
+
+        # Send mark message
+        mark_data = {
+            "event": "mark",
+            "streamSid": self._stream_sid,
+            "mark": {"name": str(self._sequence_no)},
+        }
+        mark = json.dumps(mark_data)
+        await self._ws.send_str(mark)
+        self._sequence_no = self._sequence_no + 1
+
+    async def close(self) -> None:
+        pass
+
+
+class PhoneAudioSource(audio_base.AudioSource):
+    """AudioSource that reads from the phone stream."""
+
+    def __init__(self, sample_rate=8000, channels=1):
+        super().__init__(sample_rate, channels)
+        self._queue: asyncio.Queue[bytes] = asyncio.Queue()
+
+    async def write(self, chunk: bytes) -> None:
+        decoded = g711.decode_ulaw(chunk).astype(numpy.int16)
+        await self._queue.put(decoded)
+
+    async def stream(self) -> AsyncGenerator[bytes, None]:
+        while True:
+            buf = await self._queue.get()
+            yield buf if self.enabled else b"\x00" * len(buf)
 
 
 async def testhandle(request):
@@ -21,8 +89,8 @@ async def websocket_handler(request):
     await ws.prepare(request)
     logging.info("Websocket connection ready")
 
-    source = audio_phone.PhoneAudioSource()
-    sink = audio_phone.PhoneAudioSink(ws)
+    source = PhoneAudioSource()
+    sink = PhoneAudioSink(ws)
     params = VoiceSessionParams(
         agent_id=args.agent,
         tts_voice=args.tts_voice,
@@ -65,7 +133,7 @@ async def websocket_handler(request):
                 await session.warmup()
             if data["event"] == "start":
                 logging.info(f"Received start message={msg}")
-                sink.stream_sid(data["streamSid"])
+                sink.stream_sid = data["streamSid"]
                 await session.start()
             if data["event"] == "media":
                 payload = data["media"]["payload"]
